test: rename unit tests
